requirements-airflow.txt
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-common-sql==1.7.1
psycopg2-binary==2.9.9
tqdm>=4.66.1
pandas==2.0.3
pyarrow>=12.0.0
warcio==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
tldextract==5.1.1
python-whois==0.8.0
urllib3==2.0.7
ipwhois==1.2.0 